Disable building wear app if they are not bundled.
88e4d4 added this to detect package discrepancy
during debug build, so that the detection can be done
as early as possible.

However it did this by building/checking the release
version of the wear app against the debug (or any other
build type) version of the main app. This may not work
in case of appId customization on either build type.

The proper fix will require proper varaint-aware
dependency management which we do not yet have, so
this change mostly disable the previous change
and adds some tests.

https://code.google.com/p/android/issues/detail?id=152837

(cherry picked from commit 3d224a1cca5bcaf22e5dfef9b58363a565242540)

Change-Id: I402f8d22815c2014ebb6ff3555dc001bffc08e80