Fix issue with duplicate dependencies and proguard.
When a duplicate dependency was found, we would remove
it from the package configuration, but not from the
compile one (even though since the dependencies of the
tested app are added to the classpath it was not needed
anywhere).

This had the side-effect that for proguarded tested
variant, the dependency would be added to proguard twice:
- as a tested dependency, passed as a inLibrary.
- as a provided-only dependency on the test variant, also
  passed as a inLibrary.

Proguard complains about the duplicate inclusion.

This fix completely removes duplicate dependencies from
the test app. They are not used for compilation nor
packaging.

Change-Id: If00de4b750d714e34096ea4b566b2dce83356b4c
(cherry picked from commit d61dc6a70148eaf84ba326ca3df3dc6f63cd48a2)